// src/types.ts
export enum EventType {
  CART_ABANDONED = 'cart.abandoned',
  CART_CREATED = 'cart.created',
  CART_ORDERING = 'cart.ordering',
  CART_UPDATED = 'cart.updated',
  CHANNEL_CREATED = 'channel.created',
  CHANNEL_UPDATED = 'channel.updated',
  CHANNEL_DELETED = 'channel.deleted',
  CHANNEL_ENTRIES_CREATED = 'channel.entries.created',
  CHANNEL_ENTRIES_DELETED = 'channel.entries.deleted',
  CHANNEL_ENTRIES_UPDATED = 'channel.entries.updated',
  COLLECTION_CREATED = 'collection.created',
  COLLECTION_DELETED = 'collection.deleted',
  COLLECTION_UPDATED = 'collection.updated',
  COUPON_CREATED = 'coupon.created',
  COUPON_DELETED = 'coupon.deleted',
  COUPON_UPDATED = 'coupon.updated',
  CUSTOMER_CREATED = 'customer.created',
  CUSTOMER_DELETED = 'customer.deleted',
  CUSTOMER_UPDATED = 'customer.updated',
  DEVICE_CREATED = 'device.created',
  DEVICE_DELETED = 'device.deleted',
  DEVICE_UPDATED = 'device.updated',
  ORDER_ATTACHMENTS_EXPIRED = 'order.attachments.expired',
  ORDER_ATTACHMENTS_READY = 'order.attachments.ready',
  ORDER_CANCELED = 'order.canceled',
  ORDER_CREATED = 'order.created',
  ORDER_FULFILLED = 'order.fulfilled',
  ORDER_PAID = 'order.paid',
  ORDER_REOPENED = 'order.reopened',
  ORDER_UPDATED = 'order.updated',
  PRODUCT_CREATED = 'product.created',
  PRODUCT_DELETED = 'product.deleted',
  PRODUCT_UPDATED = 'product.updated',
  ROLE_CREATED = 'role.created',
  ROLE_DELETED = 'role.deleted',
  ROLE_UPDATED = 'role.updated',
}

export enum ViewType {
  CHANNEL_ENTRIES_LIST = 'channel.entries.list',
  CHANNEL_ENTRIES_SHOW = 'channel.entries.show',
  COLLECTION_EDIT = 'collection.edit',
  COLLECTION_LIST = 'collection.list',
  COUPON_EDIT = 'coupon.edit',
  COUPON_LIST = 'coupon.list',
  CUSTOMER_LIST = 'customer.list',
  CUSTOMER_SHOW = 'customer.show',
  ORDER_LIST = 'order.list',
  ORDER_SHOW = 'order.show',
  PRODUCT_LIST = 'product.list',
  PRODUCT_SHOW = 'product.show',
}